Fix GaussianMixture crashing on torch.float64 input (fixes #2)
GMMInitializer and several internal tensor constructions hardcoded
float32 (or the process default dtype) regardless of the input data's
dtype, so fitting on float64 data crashed inside torch.cdist during
k-means++ initialization. GaussianMixture now tracks the dtype of the
data passed to fit() and uses it consistently for all parameters,
priors, and GMMInitializer outputs.

Also fixes two related bugs the same audit turned up: silhouette_score
crashed on float64 X via the same hardcoded-dtype pattern, and
plot_gmm's spherical/tied_spherical ellipses used a wrong-sized
identity matrix (crashing outright for tied_spherical, silently wrong
for spherical).

Includes tgmm/synthetic_data.py (generate_gmm_data), a synthetic-data
helper already wired into __init__.py and used throughout the tutorial
notebooks, and exposes tgmm.__version__.

+++ tgmm/metrics.py
@@ -170,8 +170,8 @@ class ClusteringMetrics:
         labels = labels.to(X.device)
         distances = torch.cdist(X, X)  # (n_samples, n_samples)
         
-        A = torch.zeros(labels.size(0), dtype=torch.float, device=X.device)
-        B = torch.full((labels.size(0),), float('inf'), dtype=torch.float, device=X.device)
+        A = torch.zeros(labels.size(0), dtype=X.dtype, device=X.device)
+        B = torch.full((labels.size(0),), float('inf'), dtype=X.dtype, device=X.device)
 
         # For each cluster i, compute intra-cluster distances and inter-cluster distances
         for i in range(n_components):
